fix(api): fix api product (#31)
